requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
black==23.10.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
DateTime==5.2
defusedxml==0.8.0rc2
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.3
django-redis==5.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
exceptiongroup==1.1.3
Faker==19.6.2
hiredis==2.2.3
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
psycopg2-binary==2.9.7
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pytest==7.4.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
ruff==0.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.6
zope.interface==6.1
